+++ Changelog.md
@@ -4,6 +4,14 @@ The change log describes what is "Added", "Removed", "Changed" or "Fixed" betwee
 
 ## UNRELEASED
 
+### Added
+
+* `ConnectException` that is thrown when you fail to connect to Redis
+
+### Fixed
+
+* If a DSN is provided to redis we make sure to select a database for you. 
+
 ## 0.3.4
 
 No changelog before this version